+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.20.0] - 2024-04-01
+
+### Added
+
+- DAP/LSP: `tools.cargo_override` option to
+  override the `cargo` command for runnables/debuggables/testables.
+
 ## [4.19.0] - 2024-04-01
 
 ### Added